Revert commits made to resolve a gdk-pixbuf+svg propagation issue.
Fixes <https://issues.guix.gnu.org/50957>.

This reverts the following commits:

 - 42231bc (gnu: cheese: Propagate gdk-pixbuf+svg rather than gdk-pixbuf)
 - ad6a950 (gnu: mutter: Propagate gdk-pixbuf+svg rather than gdk-pixbuf)
 - 5e02d1a (gnu: cogl: Propagate gdk-pixbuf+svg rather than gdk-pixbuf)
 - 5c467b2 (gnu: libgweather: Propagate gdk-pixbuf+svg rather than gdk-pixbuf)
 - 26487f6 (gnu: librsvg-next: Do not propagate gdk-pixbuf)
 - d891922 (gnu: librsvg: Do not propagate gdk-pixbuf)
 - 22476d3 (gnu: libgsf: Do not propagate gdk-pixbuf)
 - 099ca0e (gnu: libnotify: Propagate gdk-pixbuf+svg rather than gdk-pixbuf)
 - fb6a1ff (gnu: appstream-glib: Propagate gdk-pixbuf+svg rather than gdk-pixbuf)
 - 4fb525d (gnu: weasyprint: Propagate gdk-pixbuf+svg rather than gdk-pixbuf)

These commits were made to resolve a gdk-pixbuf loaders clash that could occur
in a profile when both gdk-pixbuf and gdk-pixbuf+svg were propagated.
Unfortunately, librsvg really needs to propagate gdk-pixbuf itself, which
would introduce a cycle.  Packages such as 'efl' were broken as a result of
not propagating gdk-pixbuf.

A more definitive fix will be proposed for core-updates, which includes a
profile hook to generate the gdk-pixbuf loaders.cache file, as well as a build
phase and search path specification, that should resolve the issue for good.